requirements.txt
pandas>=2.2
pyyaml
thefuzz>=0.22
xlsxwriter
jinja2